+++ examples/ai-functions/src/telemetry/stream-text.ts
@@ -2,8 +2,8 @@ import { anthropic } from '@ai-sdk/anthropic';
 import { getNodeAutoInstrumentations } from '@opentelemetry/auto-instrumentations-node';
 import { NodeSDK } from '@opentelemetry/sdk-node';
 import { ConsoleSpanExporter } from '@opentelemetry/sdk-trace-node';
-import { streamText, registerTelemetryIntegration } from 'ai';
-import { OpenTelemetryIntegration } from '@ai-sdk/otel';
+import { streamText, registerTelemetry } from 'ai';
+import { OpenTelemetry } from '@ai-sdk/otel';
 import { run } from '../lib/run';
 
 const sdk = new NodeSDK({
@@ -12,7 +12,7 @@ const sdk = new NodeSDK({
 });
 
 sdk.start();
-registerTelemetryIntegration(new OpenTelemetryIntegration());
+registerTelemetry(new OpenTelemetry());
 
 run(async () => {
   const result = streamText({
@@ -23,7 +23,7 @@ run(async () => {
       something: 'custom',
       someOtherThing: 'other-value',
     },
-    experimental_telemetry: {
+    telemetry: {
       functionId: 'my-awesome-function',
     },
   });