+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,22 @@
 
 ### Patch Changes
 
+- Updated dependencies [72cb801]
+  - ai@7.0.0-beta.110
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [ec98264]
+- Updated dependencies [eea8d98]
+- Updated dependencies [75ef93e]
+  - ai@7.0.0-beta.109
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@7.0.0-beta.108
 
 ## 0.0.1-beta.0