+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/svelte
 
+## 5.0.0-beta.110
+
+### Patch Changes
+
+- Updated dependencies [72cb801]
+  - ai@7.0.0-beta.110
+
+## 5.0.0-beta.109
+
+### Patch Changes
+
+- Updated dependencies [ec98264]
+- Updated dependencies [eea8d98]
+- Updated dependencies [75ef93e]
+  - ai@7.0.0-beta.109
+  - @ai-sdk/provider-utils@5.0.0-beta.25
+
 ## 5.0.0-beta.108
 
 ### Patch Changes